+++ bika/lims/content/analysisrequest.py
@@ -149,8 +149,6 @@ schema = BikaSchema.copy() + Schema((
             label=_("Contact"),
             render_own_label=True,
             size=20,
-            helper_js=("bika_widgets/referencewidget.js",
-                       "++resource++bika.lims.js/contact.js"),
             description=_("The primary contact of this sample, "
                           "who will receive notifications and publications "
                           "via email"),
@@ -175,10 +173,9 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'CCContact',
         multiValued=1,
-        vocabulary_display_path_bound=sys.maxsize,
         allowed_types=('Contact',),
         referenceClass=HoldingReference,
         relationship='AnalysisRequestCCContact',
@@ -227,7 +224,7 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Client',
         required=1,
         allowed_types=('Client',),
@@ -256,7 +253,7 @@ schema = BikaSchema.copy() + Schema((
     # Field for the creation of Secondary Analysis Requests.
     # This field is meant to be displayed in AR Add form only. A viewlet exists
     # to inform the user this Analysis Request is secondary
-    ReferenceField(
+    UIDReferenceField(
         "PrimaryAnalysisRequest",
         allowed_types=("AnalysisRequest",),
         referenceClass=HoldingReference,
@@ -296,7 +293,7 @@ schema = BikaSchema.copy() + Schema((
         )
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Batch',
         allowed_types=('Batch',),
         relationship='AnalysisRequestBatch',
@@ -327,13 +324,13 @@ schema = BikaSchema.copy() + Schema((
                 {'columnName': 'getClientTitle', 'width': '30',
                  'label': _('Client'), 'align': 'left'},
             ],
-            force_all = False,
+            force_all=False,
             ui_item="getId",
             showOn=True,
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'SubGroup',
         required=False,
         allowed_types=('SubGroup',),
@@ -366,7 +363,7 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Template',
         allowed_types=('ARTemplate',),
         referenceClass=HoldingReference,
@@ -376,8 +373,10 @@ schema = BikaSchema.copy() + Schema((
         write_permission=FieldEditTemplate,
         widget=ReferenceWidget(
             label=_("Sample Template"),
-            description=_("The predefined values of the Sample template are set "
-                          "in the request"),
+            description=_(
+                "The predefined values of the Sample template are set in the "
+                "request"
+            ),
             size=20,
             render_own_label=True,
             visible={
@@ -392,35 +391,11 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    # TODO Remove Profile field (in singular)
-    ReferenceField(
-        'Profile',
-        allowed_types=('AnalysisProfile',),
-        referenceClass=HoldingReference,
-        relationship='AnalysisRequestAnalysisProfile',
-        mode="rw",
-        read_permission=View,
-        write_permission=ModifyPortalContent,
-        widget=ReferenceWidget(
-            label=_("Analysis Profile"),
-            description=_("Analysis profiles apply a certain set of analyses"),
-            size=20,
-            render_own_label=True,
-            visible=False,
-            catalog_name='bika_setup_catalog',
-            base_query={"is_active": True,
-                        "sort_on": "sortable_title",
-                        "sort_order": "ascending"},
-            showOn=False,
-        ),
-    ),
-
-    ReferenceField(
+    UIDReferenceField(
         'Profiles',
         multiValued=1,
         allowed_types=('AnalysisProfile',),
         referenceClass=HoldingReference,
-        vocabulary_display_path_bound=sys.maxsize,
         relationship='AnalysisRequestAnalysisProfiles',
         mode="rw",
         read_permission=View,
@@ -584,7 +559,8 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    DateTimeField('DatePreserved',
+    DateTimeField(
+        "DatePreserved",
         mode="rw",
         read_permission=View,
         write_permission=FieldEditDatePreserved,
@@ -600,7 +576,8 @@ schema = BikaSchema.copy() + Schema((
             },
         ),
     ),
-    StringField('Preserver',
+    StringField(
+        "Preserver",
         required=0,
         mode="rw",
         read_permission=View,
@@ -618,7 +595,8 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
     # TODO Sample cleanup - This comes from partition
-    DurationField('RetentionPeriod',
+    DurationField(
+        "RetentionPeriod",
         required=0,
         mode="r",
         read_permission=View,
@@ -643,7 +621,7 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Specification',
         required=0,
         primary_bound=True,  # field changes propagate to partitions
@@ -695,7 +673,7 @@ schema = BikaSchema.copy() + Schema((
         widget=ComputedWidget(visible=False),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'PublicationSpecification',
         required=0,
         allowed_types='AnalysisSpec',
@@ -706,7 +684,8 @@ schema = BikaSchema.copy() + Schema((
         widget=ReferenceWidget(
             label=_("Publication Specification"),
             description=_(
-                "Set the specification to be used before publishing a Sample."),
+                "Set the specification to be used before publishing a Sample."
+            ),
             size=20,
             render_own_label=True,
             visible={
@@ -898,7 +877,7 @@ schema = BikaSchema.copy() + Schema((
     ),
 
     # TODO Remove - Is this still necessary?
-    ReferenceField(
+    UIDReferenceField(
         'DefaultContainerType',
         allowed_types=('ContainerType',),
         relationship='AnalysisRequestContainerType',
@@ -970,7 +949,7 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Attachment',
         multiValued=1,
         allowed_types=('Attachment',),
@@ -1005,9 +984,8 @@ schema = BikaSchema.copy() + Schema((
         )
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Invoice',
-        vocabulary_display_path_bound=sys.maxsize,
         allowed_types=('Invoice',),
         referenceClass=HoldingReference,
         relationship='AnalysisRequestInvoice',
@@ -1246,7 +1224,7 @@ schema = BikaSchema.copy() + Schema((
     ),
 
     # The Primary Sample the current sample was detached from
-    ReferenceField(
+    UIDReferenceField(
         "DetachedFrom",
         allowed_types=("AnalysisRequest",),
         relationship="AnalysisRequestDetachedFrom",
@@ -1261,7 +1239,7 @@ schema = BikaSchema.copy() + Schema((
 
     # The Analysis Request the current Analysis Request comes from because of
     # an invalidation of the former
-    ReferenceField(
+    UIDReferenceField(
         'Invalidated',
         allowed_types=('AnalysisRequest',),
         relationship='AnalysisRequestRetracted',
@@ -1272,14 +1250,6 @@ schema = BikaSchema.copy() + Schema((
         widget=ReferenceWidget(
             visible=False,
         ),
-    ),
-
-    # The Analysis Request that was automatically generated due to the
-    # invalidation of the current Analysis Request
-    ComputedField(
-        'Retest',
-        expression="here.get_retest()",
-        widget=ComputedWidget(visible=False)
     ),
 
     # For comments or results interpretation
@@ -1350,7 +1320,7 @@ schema = BikaSchema.copy() + Schema((
                           "clients."),
             format="radio",
             render_own_label=True,
-            visible={'add': 'edit',}
+            visible={'add': 'edit'}
         ),
     ),
 )
@@ -1669,7 +1639,7 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
         # Keywords of the contained services
         billable_service_keys = map(
             lambda s: s.getKeyword(), set(billable_profile_services))
-        # The billable items contain billable profiles and single selected analyses
+        # Billable items contain billable profiles and single selected analyses
         billable_items = billable_profiles
         # Get the analyses to be billed
         exclude_rs = ["retracted", "rejected"]
@@ -2002,10 +1972,10 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
             sets = [adv] if 'hidden' in adv else []
 
         # Created by using an AR Profile?
-        if not sets and self.getProfiles():
-            adv = []
-            adv += [profile.getAnalysisServiceSettings(uid) for profile in
-                    self.getProfiles()]
+        profiles = self.getProfiles()
+        if not sets and profiles:
+            adv = [profile.getAnalysisServiceSettings(uid) for profile in
+                   profiles]
             sets = adv if 'hidden' in adv[0] else []
 
         return sets[0] if sets else {'uid': uid}
@@ -2151,15 +2121,27 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
                     "It can not hold an own value!")
         return None
 
-    def get_retest(self):
-        """Returns the Analysis Request automatically generated because of the
-        retraction of the current analysis request
+    def getRawRetest(self):
+        """Returns the UID of the Analysis Request that has been generated
+        automatically because of the retraction of the current Analysis Request
         """
-        relationship = "AnalysisRequestRetracted"
-        retest = self.getBackReferences(relationship=relationship)
-        if retest and len(retest) > 1:
-            logger.warn("More than one retest for {0}".format(self.getId()))
-        return retest and retest[0] or None
+        relationship = self.getField("Invalidated").relationship
+        uids = get_backreferences(self, relationship=relationship)
+        return uids[0] if uids else None
+
+    @deprecated("Use getRetest instead")
+    def get_retest(self):
+        """Returns the Analysis Request that has been generated automatically
+        because of the retraction of the current Analysis Request
+        """
+        return self.getRetest()
+
+    def getRetest(self):
+        """Returns the Analysis Request that has been generated automatically
+        because of the retraction of the current Analysis Request
+        """
+        uid = self.getRawRetest()
+        return api.get_object_by_uid(uid, default=None)
 
     def getAncestors(self, all_ancestors=True):
         """Returns the ancestor(s) of this Analysis Request
@@ -2208,8 +2190,8 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
 
         This method is used as metadata
         """
-        rel_id = "AnalysisRequestParentAnalysisRequest"
-        return get_backreferences(self, relationship=rel_id)
+        relationship = self.getField("ParentAnalysisRequest").relationship
+        return get_backreferences(self, relationship=relationship)
 
     def isPartition(self):
         """Returns true if this Analysis Request is a partition
@@ -2246,11 +2228,19 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
             parent = self.getParentAnalysisRequest()
             alsoProvides(parent, IAnalysisRequestWithPartitions)
 
+    def getRawSecondaryAnalysisRequests(self):
+        """Returns the UIDs of the secondary Analysis Requests from this
+        Analysis Request
+        """
+        relationship = self.getField("PrimaryAnalysisRequest").relationship
+        return get_backreferences(self, relationship)
+
     def getSecondaryAnalysisRequests(self):
         """Returns the secondary analysis requests from this analysis request
         """
-        relationship = "AnalysisRequestPrimaryAnalysisRequest"
-        return self.getBackReferences(relationship=relationship)
+        uids = self.getRawSecondaryAnalysisRequests()
+        uc = api.get_tool("uid_catalog")
+        return [api.get_object(brain) for brain in uc(UID=uids)]
 
     def setDateReceived(self, value):
         """Sets the date received to this analysis request and to secondary
@@ -2394,10 +2384,10 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
 
         # convert relative URLs to absolute URLs
         # N.B. This is actually a TinyMCE issue, but hardcoded in Plone:
-        #      https://www.tiny.cloud/docs/configure/url-handling/#relative_urls
+        # https://www.tiny.cloud/docs/configure/url-handling/#relative_urls
         image_sources = re.findall(IMG_SRC_RX, html)
 
-        # we need a trailing slash so that urljoin does not remove the last segment
+        # add a trailing slash so that urljoin doesn't remove the last segment
         base_url = "{}/".format(api.get_url(self))
 
         for src in image_sources: