+++ bika/lims/exportimport/setupdata/__init__.py
@@ -2292,13 +2292,13 @@ class Analysis_Requests(WorksheetImporter):
                              getFullname=row['CCContact_Fullname'])[0].getObject()
                 obj.setCCContact(contact)
             if row['AnalysisProfile_title']:
-                profile = pc(portal_type="AnalysisProfile",
-                             title=row['AnalysisProfile_title'].getObject())
-                obj.setProfile(profile)
+                profiles = pc(portal_type="AnalysisProfile",
+                              title=row['AnalysisProfile_title'])[0].getObject()
+                obj.setProfiles([profiles])
             if row['ARTemplate_title']:
                 template = pc(portal_type="ARTemplate",
-                             title=row['ARTemplate_title'])[0].getObject()
-                obj.setProfile(template)
+                              title=row['ARTemplate_title'])[0].getObject()
+                obj.setTemplate(template)
 
             obj.unmarkCreationFlag()